All test pass again !!!!! Loggin the bubbles is all broken, however... But not to worry, I should get around to fixing it pretty soon
